HD Tanks are not ready for sending to ERP!!!!
This is an intermediate commit to modify the existing definition of all tank parts made at LTI.  Original BOM behaved as if parts were cut at TACTIC.  This affected all parts referencing material 9331250, which now is mill finish instead of #4 finish.  -1 and -2 of that part are for #4 finish on one and both sides respectively. 9331000-A was added to identify the external processing cost of converting mill finish to #4 both sides.